Move analyzer loaders out of package root into manager.py
