+++ anchore_engine/configuration/localconfig.py
@@ -101,6 +101,17 @@ default_required_config_params = {
 CRED_CACHE_TTL = int(os.getenv("ANCHORE_INTERNAL_CRED_CACHE_TTL", 600))
 CRED_CACHE_LOCK_WAIT_SEC = int(os.getenv("ANCHORE_INTERNAL_CRED_CACHE_WAIT_SEC", 3))
 
+ANALYZER_SEARCH_PATHS = ["anchore_engine.analyzers"]
+
+
+def register_analyzers(module_path):
+    global ANALYZER_SEARCH_PATHS
+    ANALYZER_SEARCH_PATHS.append(module_path)
+
+
+def analyzer_paths():
+    return ANALYZER_SEARCH_PATHS
+
 
 def update_merge(base, override):
     if not isinstance(base, dict) or not isinstance(override, dict):
@@ -321,21 +332,7 @@ def load_config(configdir=None, configfile=None, validate_params=None):
         if ext_config.get("metadata_types", []):
             localconfig["image_metadata_types"].extend(ext_config.get("metadata_types"))
 
-    # any special overrides/deprecation handling here
-    try:
-        analyzer_config = localconfig.get("services", {}).get("analyzer", {})
-        if (
-            analyzer_config
-            and analyzer_config.get("analyzer_driver", "localanchore") != "nodocker"
-        ):
-            if not os.path.exists("/usr/bin/anchore"):
-                logger.warn(
-                    "the 'localanchore' analyzer driver has been removed from anchore-engine - defaulting to 'nodocker' analyzer driver"
-                )
-                localconfig["services"]["analyzer"]["analyzer_driver"] = "nodocker"
-    except Exception as err:
-        logger.warn(str(err))
-        pass
+    localconfig["services"]["analyzer"]["analyzer_driver"] = "nodocker"
 
     return localconfig
